@@ -22,13 +22,32 @@ public class splash extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_splash);
 
-        logo=findViewById(R.id.logoimg);
-        name=findViewById(R.id.logonameimg);
-        own1=findViewById(R.id.ownone);
-        own2=findViewById(R.id.owntwo);
+        //notification 24/4/2024
+//        if (getIntent().getExtras() != null) {
+//            String userId=getIntent().getExtras().getString("userId");
+//            FirebaseUtil.allChatroomCollectionReference().document(userId).get()
+//                    .addOnCompleteListener(task -> {
+//                        if(task.isSuccessful()){
+//                            //copy from useradapter
+//                            Intent intent=new Intent(splash.this,MainActivity.class);
+//                            intent.setFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION);
+//                            startActivity(intent);
+//
+//
+//                        }
+//
+//                    });
+//
+//        }
+     // else {
 
-        topAnim= AnimationUtils.loadAnimation(this,R.anim.top_animation);
-        bottomAnim= AnimationUtils.loadAnimation(this,R.anim.bottom_animation);
+        logo = findViewById(R.id.logoimg);
+        name = findViewById(R.id.logonameimg);
+        own1 = findViewById(R.id.ownone);
+        own2 = findViewById(R.id.owntwo);
+
+        topAnim = AnimationUtils.loadAnimation(this, R.anim.top_animation);
+        bottomAnim = AnimationUtils.loadAnimation(this, R.anim.bottom_animation);
 
         logo.setAnimation(topAnim);
         name.setAnimation(bottomAnim);
@@ -39,10 +58,12 @@ public class splash extends AppCompatActivity {
         new Handler().postDelayed(new Runnable() {
             @Override
             public void run() {
-                Intent intent=new Intent(splash.this, MainActivity.class);
+                Intent intent = new Intent(splash.this, MainActivity.class);
                 startActivity(intent);
                 finish();
             }
-        },4000);
+        }, 4000);
+
+      // }
     }
 }
